Add new yarn script `clean-view-custom` to be called at the beginning of all `primo-explore-devenv:run:*` scripts [+]
The `run` gulp task in `primo-explore-devenv` does not stop the local Primo server if the JS build fails.  We will attempt to fix that in `primo-explore-devenv`, but for now, at least clean the build to reduce risk of a failed build from causing confusion.

For now we are just cleaning the JS build.  We will eventually want to clean CSS build as well, but we currently have no source CSS files to re-create the empty custom1.css file.  So if we delete it it will not be rebuilt, and the local Primo will fail to start without it.

Note that this clean step does not run when a file is edited and gulp reloads.  It only runs when the `yarn` scripts are run, or when the user runs the script/clean-view-custom.sh script directly.